fix: remove Copy from Material
i experimented with having Hit hold a reference to a Material instead of
cloning the Material. this was significantly slower than cloning the
Material. my guess is that going through a reference to the heap is not
cache-friendly, while cloning a 40 byte struct is sufficiently cheap and
also maintains locality for the scatter computation. neat.