feat: rework SAH partitioning implementation and add a per-surface strategy
im happier with this implementation since it:
- avoids repeatedly shuffling surfaces around in memory
- avoids repeatedly computing intermediate bounding boxes for partitions
- separates selecting candidates from selecting the best split

in some sense, it wasn't a "good" idea to write this, since it basically
doesn't matter for any scenes im rendering now or even plan to render in
the future. it brought me some satisfaction though.

i think this does the computation in a smarter way. i might take a pass
at cleaning this up another time.