feat: implement fork_union rendering function
anecdotally, it is slower or almost as fast as the rayon implementation.
seems a little odd? the project's benchmarks led me to expect a
performance improvement.

i wonder why. maybe my implementation is wrong somewhere. the shape of
the workload isn't exactly ideal for fork_union i would guess though.
individual rays can vary wildly in the number of intersections and that
depth can't be perfectly predicted. individual pixels are the same, and
since i'm parallelizing over pixels, i went with the dynamic operation
and accepted some of its costs.

anyway, something to explore perhaps